refactor: move push-notifications model to features from entities
