+++ apps/web/src/_pages/auth/sign-in/EmailLoginPage.tsx
@@ -14,7 +14,7 @@ import { AuthContainer } from '@/widgets/auth';
 import { useSignInMutation } from '@/features/auth';
 import { usePushNotification } from '@/features/notification';
 
-import { SigninRequestDto, signInSchema } from '@/entities/auth';
+import { type SigninRequestDto, signInSchema } from '@/entities/auth';
 
 import { toast } from '@/shared/model';
 import { TokenManager } from '@/shared/storage';
