+++ apps/web/src/entities/notification/api/index.ts
@@ -1,1 +1,2 @@
 export * from './get';
+export * from './post';
