+++ apps/web/src/entities/notification/index.ts
@@ -1,2 +1,3 @@
 export * from './model';
 export * from './lib';
+export * from './api';
