+++ apps/web/src/features/notification/model/index.ts
@@ -1,2 +1,2 @@
-export * from './mutations';
 export * from './hooks';
+export * from './mutations';
